+++ Robust.Shared/Physics/B2DynamicTree.cs
@@ -583,6 +583,9 @@ namespace Robust.Shared.Physics
                 indexNode.Height = Math.Max(child1Node.Height, child2Node.Height) + 1;
                 indexNode.Aabb = child1Node.Aabb.Union(child2Node.Aabb);
 
+                if (index == indexNode.Parent)
+                    throw new Exception($"Infinite loop in B2DynamicTree.Balance(). Trace: {Environment.StackTrace}");
+
                 index = indexNode.Parent;
             }
 
@@ -1111,7 +1114,6 @@ namespace Robust.Shared.Physics
             Debugger.Break();
             throw new InvalidOperationException(msg);
         }
-
 
         private IEnumerable<(Proxy, Node)> DebugAllocatedNodesEnumerable
         {